requirements.txt
pandas
yfinance
streamlit
openai
python-dotenv
tqdm
pydantic
